Develop (#10)
* Designing home page

* Designing home page more

* changes in sqlite

* create user section

* add registration form

* add registration templates

* add register views

* add register views functionality

* add email field in forms

* improving form design

* improving final details on form layout

* add login url,views,template

* add login and redirect url

* add logout functionality

* add small design change

* add header restriction

* add header another restriction

* add page restrictions

* add foreign key

* update task input fields

* update task filter in user only

* add security

* postgres conf

+++ taskmate/todolist_app/views.py
@@ -4,22 +4,27 @@ from todolist_app.models import TaskList
 from todolist_app.forms import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 
+@login_required
 def todolist(request):
     if request.method=="POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
         messages.success(request,("New Task Added!"))    
         return redirect('todolist')    
     # this is the get request    
     else:    
-        all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manage= request.user)
         paginator = Paginator(all_tasks, 5)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
         
         return render(request, 'todolist.html',{'all_tasks':all_tasks})
+
 
 def contact(request):
     context={
@@ -42,11 +47,17 @@ def index(request):
     }
     return render(request, 'index.html',context)
 
+@login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed."))
+         
     return redirect('todolist')
 
+@login_required
 def edit_task(request, task_id):
     if request.method=="POST":
         task = TaskList.objects.get(pk=task_id)
@@ -61,13 +72,17 @@ def edit_task(request, task_id):
         
         return render(request, 'edit.html',{'task_obj':task_obj})
     
-    
+@login_required    
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed."))    
     return redirect('todolist')
 
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     task.done = False
